fix(twig): organize utils
